requirements.txt
Flask
pymongo[srv]
requests
python-dotenv
bcrypt
pytest
pytest-flask
mongomock
pytest-playwright
